refactor (AuthController, AuthService): rename package where the AuthController and AuthService are saved to
